Oficina de Integracao III/Backend/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
cudatoolkit==11.2.2
cudnn==8.1.0.77
tensorflow==2.6.2
numpy==1.19.5
psycopg2-binary
bcrypt
pyjwt
silence-tensorflow